requirements.txt
face-recognition===1.3.0
face_recognition_models===0.3.0
Flask===3.1.0
numpy===1.26.4
opencv-python===4.8.0.76
psycopg2===2.9.10
python-dotenv
requests
